streamlit_dashboard/requirements.txt
streamlit==1.31.0
snowflake-connector-python==3.7.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
